refactor: transaction consumer로 책임 분리하기 Fixes #11
